refactor(playground): isolate Signals chart state
Move Sankey measurement and snapshot playback behavior into focused hooks, and separate the timeline UI from the flow container so each concern can be tested independently.\n\nAddresses https://github.com/mastra-ai/mastra/pull/19871#discussion_r3631813507\nAddresses https://github.com/mastra-ai/mastra/pull/19871#discussion_r3631814863\nAddresses https://github.com/mastra-ai/mastra/pull/19871#discussion_r3631825952\nAddresses https://github.com/mastra-ai/mastra/pull/19871#discussion_r3631828243
