+++ packages/playground/src/ee/signals/sankey-signals.tsx
@@ -1,12 +1,10 @@
-import { Button } from '@mastra/playground-ui/components/Button';
 import { Card, CardContent, CardFooter, CardHeader } from '@mastra/playground-ui/components/Card';
 import { nodeColor, Sankey, SankeyChart } from '@mastra/playground-ui/components/SankeyChart';
 import type { SankeyChartColumn, SankeyChartRecord } from '@mastra/playground-ui/components/SankeyChart';
-import { Slider } from '@mastra/playground-ui/components/Slider';
 import { getSignalHue, SignalsOverviewPage as SignalsEmptyState } from '@mastra/playground-ui/ee/signals';
-import { Pause, Play } from 'lucide-react';
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo } from 'react';
 
+import { useSnapshotPlayback } from './hooks/use-snapshot-playback';
 import { useThemeFlows } from './hooks/use-theme-flows';
 import { useThemeSnapshots } from './hooks/use-theme-snapshots';
 import {
@@ -16,9 +14,11 @@ import {
   getSignalRecordNodeValue,
   stabilizeThemeFlow,
 } from './sankey-signals-data';
+import { formatSignalName, formatSnapshotWindow, traceLabel } from './signal-formatting';
 import { SignalsErrorState } from './signals-error-state';
 import { SignalsLoadingSkeleton } from './signals-loading-skeleton';
-import type { ThemeFlowResponse, ThemeNode, ThemeSnapshot, TraceSignalName } from './types';
+import { SnapshotTimeline } from './snapshot-timeline';
+import type { ThemeFlowResponse, ThemeNode, TraceSignalName } from './types';
 import { Link } from '@/lib/link';
 
 export interface SankeySignalsProps {
@@ -26,86 +26,6 @@ export interface SankeySignalsProps {
   entityType?: string;
   signalNames: TraceSignalName[];
   height?: number;
-}
-
-function formatSignalName(signalName: TraceSignalName) {
-  return signalName.charAt(0).toUpperCase() + signalName.slice(1);
-}
-
-function traceLabel(count: number) {
-  return `${count} ${count === 1 ? 'trace' : 'traces'}`;
-}
-
-function formatSnapshotWindow(startedAt: string, endedAt: string) {
-  const start = new Date(startedAt);
-  const end = new Date(endedAt);
-  const monthDay = new Intl.DateTimeFormat('en-US', { month: 'short', day: 'numeric', timeZone: 'UTC' });
-  const day = new Intl.DateTimeFormat('en-US', { day: 'numeric', timeZone: 'UTC' });
-  const year = new Intl.DateTimeFormat('en-US', { year: 'numeric', timeZone: 'UTC' });
-  const sameYear = start.getUTCFullYear() === end.getUTCFullYear();
-  const sameMonth = sameYear && start.getUTCMonth() === end.getUTCMonth();
-  const sameDay = sameMonth && start.getUTCDate() === end.getUTCDate();
-
-  if (sameDay) return `${monthDay.format(start)}, ${year.format(start)}`;
-  if (sameMonth) return `${monthDay.format(start)}–${day.format(end)}, ${year.format(end)}`;
-  if (sameYear) return `${monthDay.format(start)}–${monthDay.format(end)}, ${year.format(end)}`;
-  return `${monthDay.format(start)}, ${year.format(start)}–${monthDay.format(end)}, ${year.format(end)}`;
-}
-
-function SnapshotTimeline({
-  snapshots,
-  selectedIndex,
-  isPlaying,
-  onPlayingChange,
-  onSnapshotChange,
-}: {
-  snapshots: ThemeSnapshot[];
-  selectedIndex: number;
-  isPlaying: boolean;
-  onPlayingChange: (isPlaying: boolean) => void;
-  onSnapshotChange: (index: number) => void;
-}) {
-  const snapshot = snapshots[selectedIndex];
-
-  if (!snapshot) return null;
-
-  return (
-    <section
-      aria-label="Snapshot timeline"
-      className="flex flex-wrap items-center gap-3 rounded-lg border border-border1 bg-surface2 px-3 py-2.5 sm:px-4"
-    >
-      {snapshots.length > 1 ? (
-        <>
-          <Button
-            aria-label={isPlaying ? 'Pause snapshots' : 'Play snapshots'}
-            onClick={() => onPlayingChange(!isPlaying)}
-            size="sm"
-            type="button"
-            variant="outline"
-          >
-            {isPlaying ? <Pause aria-hidden="true" /> : <Play aria-hidden="true" />}
-            {isPlaying ? 'Pause' : 'Play'}
-          </Button>
-          <Slider
-            aria-label="Snapshot"
-            className="min-w-32 flex-1"
-            max={snapshots.length - 1}
-            min={0}
-            onValueChange={value => onSnapshotChange(value[0] ?? 0)}
-            step={1}
-            value={[selectedIndex]}
-          />
-        </>
-      ) : null}
-      <p
-        aria-live="polite"
-        className="w-full text-left font-mono text-xs tabular-nums text-neutral4 md:ml-auto md:w-auto md:min-w-72 md:text-right"
-      >
-        Snapshot {snapshot.ordinal}/{snapshot.total} · {formatSnapshotWindow(snapshot.startedAt, snapshot.endedAt)} ·{' '}
-        {traceLabel(snapshot.traceCount)}
-      </p>
-    </section>
-  );
 }
 
 function SignalDistribution({
@@ -256,35 +176,24 @@ function FlowCard({
 export function SankeySignals({ entityId, entityType = 'agent', signalNames, height }: SankeySignalsProps) {
   const snapshotsQuery = useThemeSnapshots(entityId, entityType, signalNames);
   const snapshots = [...(snapshotsQuery.data?.snapshots ?? [])].sort((left, right) => left.ordinal - right.ordinal);
-  const [selectedSnapshotId, setSelectedSnapshotId] = useState<string>();
-  const [isPlaying, setIsPlaying] = useState(false);
-  const matchedSnapshotIndex = snapshots.findIndex(snapshot => snapshot.snapshotId === selectedSnapshotId);
-  const selectedSnapshotIndex = matchedSnapshotIndex >= 0 ? matchedSnapshotIndex : snapshots.length - 1;
-  const snapshot = snapshots[selectedSnapshotIndex];
-  const selectSnapshot = (index: number) => setSelectedSnapshotId(snapshots[index]?.snapshotId);
-
-  const nextSnapshotId = snapshots[(selectedSnapshotIndex + 1) % snapshots.length]?.snapshotId;
   const flowQueries = useThemeFlows(
     entityId,
     entityType,
     signalNames,
     snapshots.map(candidate => candidate.snapshotId),
   );
-  const flowQuery = flowQueries[selectedSnapshotIndex];
   const isFlowPending = flowQueries.some(query => query.isPending);
   const hasFlowError = flowQueries.some(query => query.isError);
-  const flow = flowQuery?.data;
+  const { isPlaying, selectedSnapshotIndex, selectSnapshot, setIsPlaying, snapshot } = useSnapshotPlayback(
+    snapshots,
+    isFlowPending || hasFlowError,
+  );
+  const flow = flowQueries[selectedSnapshotIndex]?.data;
   const windowFlows = useMemo(() => flowQueries.flatMap(query => (query.data ? [query.data] : [])), [flowQueries]);
   const graphSummary = useMemo(
     () => (flow ? buildSignalGraphSummary(stabilizeThemeFlow(flow, windowFlows)) : undefined),
     [flow, windowFlows],
   );
-
-  useEffect(() => {
-    if (!isPlaying || snapshots.length < 2 || isFlowPending || hasFlowError) return;
-    const timer = window.setTimeout(() => setSelectedSnapshotId(nextSnapshotId), 900);
-    return () => window.clearTimeout(timer);
-  }, [hasFlowError, isFlowPending, isPlaying, nextSnapshotId, snapshots.length]);
 
   if (snapshotsQuery.isPending) return <SignalsLoadingSkeleton />;
